+++ scripts/variables.sh
@@ -8,3 +8,6 @@ clean_key_option="@tpm-clean"
 default_clean_key="M-u"
 
 SUPPORTED_TMUX_VERSION="1.9"
+
+DEFAULT_TPM_ENV_VAR_NAME="TMUX_PLUGIN_MANAGER_PATH"
+DEFAULT_TPM_PATH="$HOME/.tmux/plugins/"