+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - refactor `shared_set_tpm_path_constant` function
+- move all instructions to `docs/` dir
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings