+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - improved test runner function
 - switch to using [tmux-test](https://github.com/tmux-plugins/tmux-test)
   framework
+- add `bin/update_plugins` cli executable script
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings