+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - switch to using [tmux-test](https://github.com/tmux-plugins/tmux-test)
   framework
 - add `bin/update_plugins` cli executable script
+- refactor test `expect` scripts, make them simpler and ensure they properly
+  assert expectations
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings