+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - add `bin/update_plugins` cli executable script
 - refactor test `expect` scripts, make them simpler and ensure they properly
   assert expectations
+- refactor code that sets 'TMUX_PLUGIN_MANAGER_PATH' global env var
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings